refactor(embeddings): extract embeddings-store module
